角标、红点方法设置
新增三个方法：
/**
 *  角标设置(注意： 1、显示的话，对应位置上的dataSource必须支持显示角标，否则就算设置yes，也不会显示； 2、如果是要在现有角标数据上减一个数值，则直接传一个负值即可，如：现在角标数值是10，你需要其减1，则直接传-1即可；3、如果是需要递增，则调用cornerValueIncreasing方法)
 *
 *  @param value 角标数值
 *  @param index 对应的位置
 */
- (void)cornerValue:(int)value atIndex:(NSInteger)index;

/**
 *  角标设置，递增值(注意： 1、显示的话，对应位置上的dataSource必须支持显示角标，否则就算设置yes，也不会显示； 2、此方法为在现有角标数值的基础上增加一个数值，如：现在角标数值是10，你需要在其基础上加2，则调用此方法; 3、如果设置的数值小于0则不执行操作)
 *
 *  @param value 递增值
 *  @param index 对应的位置
 */
- (void)cornerValueIncreasing:(NSInteger)value atIndex:(NSInteger)index;

/**
 *  新消息红点(注意： 显示的话，对应位置上的dataSource必须支持显示红点，否则就算设置yes，也不会显示)
 *
 *  @param isShow 是否显示，YES为显示，NO为不显示
 *  @param index  对应的文职
 */
- (void)isNewDotShow:(BOOL)isShow atIndex:(NSInteger)index;

+++ CNCenterButtonTabBar/CNTabBar/CNTabBarButtonDataSourceModel.h
@@ -41,7 +41,7 @@
 // 是否角标数值为0时，是否隐藏
 @property (nonatomic, assign) BOOL isCornerZeroHidden;
 // 设置角标数值
-@property (nonatomic, assign) int cornerValue;
+@property (nonatomic, assign) NSInteger cornerValue;
 
 // 当前是否选中(默认为NO)
 @property (nonatomic, assign) BOOL isSelected;
